+++ src/aks-preview/azext_aks_preview/managed_cluster_decorator.py
@@ -14,10 +14,14 @@ from azure.cli.command_modules.acs._consts import (
 from azure.cli.command_modules.acs._helpers import (
     check_is_msi_cluster,
     format_parameter_name_to_option_name,
+    safe_list_get,
     safe_lower,
 )
 from azure.cli.command_modules.acs._validators import (
     extract_comma_separated_string,
+)
+from azext_aks_preview.azuremonitorprofile import (
+    ensure_azure_monitor_profile_prerequisites
 )
 from azure.cli.command_modules.acs.managed_cluster_decorator import (
     AKSManagedClusterContext,
@@ -44,8 +48,14 @@ from azext_aks_preview._consts import (
     CONST_AZURE_KEYVAULT_NETWORK_ACCESS_PRIVATE,
     CONST_AZURE_KEYVAULT_NETWORK_ACCESS_PUBLIC,
     CONST_LOAD_BALANCER_SKU_BASIC,
+    CONST_OUTBOUND_TYPE_LOAD_BALANCER,
+    CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY,
+    CONST_OUTBOUND_MIGRATION_MULTIZONE_TO_NATGATEWAY_MSG,
     CONST_PRIVATE_DNS_ZONE_NONE,
     CONST_PRIVATE_DNS_ZONE_SYSTEM,
+    CONST_EBPF_DATAPLANE_CILIUM,
+    CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY,
+    CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
 )
 from azext_aks_preview._helpers import (
     get_cluster_snapshot_by_snapshot_id,
@@ -56,6 +66,8 @@ from azext_aks_preview._loadbalancer import create_load_balancer_profile
 from azext_aks_preview._loadbalancer import (
     update_load_balancer_profile as _update_load_balancer_profile,
 )
+from azext_aks_preview._natgateway import update_nat_gateway_profile as _update_nat_gateway_profile
+
 from azext_aks_preview._podidentity import (
     _fill_defaults_for_pod_identity_profile,
     _is_pod_identity_addon_enabled,
@@ -71,6 +83,7 @@ logger = get_logger(__name__)
 
 # type variables
 ContainerServiceClient = TypeVar("ContainerServiceClient")
+ContainerServiceNetworkProfileKubeProxyConfig = TypeVar("ContainerServiceNetworkProfileKubeProxyConfig")
 ManagedCluster = TypeVar("ManagedCluster")
 ManagedClusterAddonProfile = TypeVar("ManagedClusterAddonProfile")
 ManagedClusterHTTPProxyConfig = TypeVar("ManagedClusterHTTPProxyConfig")
@@ -86,6 +99,7 @@ ManagedClusterIngressProfileWebAppRouting = TypeVar("ManagedClusterIngressProfil
 ManagedClusterSecurityProfileDefender = TypeVar("ManagedClusterSecurityProfileDefender")
 ManagedClusterSecurityProfileNodeRestriction = TypeVar("ManagedClusterSecurityProfileNodeRestriction")
 ManagedClusterWorkloadProfileVerticalPodAutoscaler = TypeVar("ManagedClusterWorkloadProfileVerticalPodAutoscaler")
+ManagedClusterLoadBalancerProfile = TypeVar("ManagedClusterLoadBalancerProfile")
 
 
 # pylint: disable=too-few-public-methods
@@ -264,6 +278,119 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         # this parameter does not need validation
         return ip_families
 
+    def _get_outbound_type(
+        self,
+        enable_validation: bool = False,
+        read_only: bool = False,
+        load_balancer_profile: ManagedClusterLoadBalancerProfile = None,
+    ) -> Union[str, None]:
+        """Internal function to dynamically obtain the value of outbound_type according to the context.
+
+        Note: All the external parameters involved in the validation are not verified in their own getters.
+
+        When outbound_type is not assigned, dynamic completion will be triggerd. By default, the value is set to
+        CONST_OUTBOUND_TYPE_LOAD_BALANCER.
+
+        This function supports the option of enable_validation. When enabled, if the value of outbound_type is
+        CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY, CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY or
+        CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING, the following checks will be performed. If load_balancer_sku is set
+        to basic, an InvalidArgumentValueError will be raised. If vnet_subnet_id is not assigned,
+        a RequiredArgumentMissingError will be raised. If any of load_balancer_managed_outbound_ip_count,
+        load_balancer_outbound_ips or load_balancer_outbound_ip_prefixes is assigned, a MutuallyExclusiveArgumentError
+        will be raised.
+        This function supports the option of read_only. When enabled, it will skip dynamic completion and validation.
+        This function supports the option of load_balancer_profile, if provided, when verifying loadbalancer-related
+        parameters, the value in load_balancer_profile will be used for validation.
+
+        :return: string or None
+        """
+        # read the original value passed by the command
+        outbound_type = self.raw_param.get("outbound_type")
+        # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
+        read_from_mc = False
+        if self.decorator_mode == DecoratorMode.CREATE:
+            if (
+                self.mc and
+                self.mc.network_profile and
+                self.mc.network_profile.outbound_type is not None
+            ):
+                outbound_type = self.mc.network_profile.outbound_type
+                read_from_mc = True
+
+        # skip dynamic completion & validation if option read_only is specified
+        if read_only:
+            return outbound_type
+
+        # dynamic completion
+        if (
+            self.decorator_mode == DecoratorMode.CREATE and
+            not read_from_mc and
+            outbound_type != CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY and
+            outbound_type != CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY and
+            outbound_type != CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING
+        ):
+            outbound_type = CONST_OUTBOUND_TYPE_LOAD_BALANCER
+
+        # validation
+        # Note: The parameters involved in the validation are not verified in their own getters.
+        if enable_validation:
+            if outbound_type in [
+                CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
+                CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY,
+                CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY,
+            ]:
+                if safe_lower(self._get_load_balancer_sku(enable_validation=False)) == CONST_LOAD_BALANCER_SKU_BASIC:
+                    raise InvalidArgumentValueError(
+                        f"{outbound_type} doesn't support basic load balancer sku"
+                    )
+
+                if outbound_type in [
+                    CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
+                    CONST_OUTBOUND_TYPE_USER_ASSIGNED_NAT_GATEWAY,
+                ]:
+                    if self.get_vnet_subnet_id() in ["", None]:
+                        raise RequiredArgumentMissingError(
+                            "--vnet-subnet-id must be specified for userDefinedRouting and it must "
+                            "be pre-configured with a route table with egress rules"
+                        )
+
+                if outbound_type == CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING:
+                    if load_balancer_profile:
+                        if (
+                            load_balancer_profile.managed_outbound_i_ps or
+                            load_balancer_profile.outbound_i_ps or
+                            load_balancer_profile.outbound_ip_prefixes
+                        ):
+                            raise MutuallyExclusiveArgumentError(
+                                "userDefinedRouting doesn't support customizing \
+                                a standard load balancer with IP addresses"
+                            )
+                    else:
+                        if (
+                            self.get_load_balancer_managed_outbound_ip_count() or
+                            self.get_load_balancer_outbound_ips() or
+                            self.get_load_balancer_outbound_ip_prefixes()
+                        ):
+                            raise MutuallyExclusiveArgumentError(
+                                "userDefinedRouting doesn't support customizing \
+                                a standard load balancer with IP addresses"
+                            )
+            if self.decorator_mode == DecoratorMode.UPDATE:
+                if outbound_type == CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY:
+                    if self.mc.agent_pool_profiles is not None and len(self.mc.agent_pool_profiles) > 1:
+                        multizoned = False
+                        for ap in self.mc.agent_pool_profiles:
+                            if ap.availability_zones:
+                                multizoned = True
+                        msg = (
+                            "\nWarning: this AKS cluster has multi-zonal nodepools, but NAT Gateway is not "
+                            "currently zone redundant. Migrating outbound connectivity to NAT Gateway could lead to "
+                            "a reduction in zone redundancy for this cluster. Continue?"
+                        )
+                        if multizoned and not self.get_yes() and not prompt_y_n(msg, default="n"):
+                            raise DecoratorEarlyExitException()
+        return outbound_type
+
     def get_load_balancer_managed_outbound_ip_count(self) -> Union[int, None]:
         """Obtain the value of load_balancer_managed_outbound_ip_count.
 
@@ -317,6 +444,13 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         """
         return self.raw_param.get('network_plugin_mode')
 
+    def get_enable_cilium_dataplane(self) -> bool:
+        """Get the value of enable_cilium_dataplane
+
+        :return: bool
+        """
+        return bool(self.raw_param.get('enable_cilium_dataplane'))
+
     def get_load_balancer_managed_outbound_ipv6_count(self) -> Union[int, None]:
         """Obtain the expected count of IPv6 managed outbound IPs.
 
@@ -369,6 +503,39 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         # this parameter does not need dynamic completion
         # this parameter does not need validation
         return load_balancer_backend_pool_type
+
+    def get_kube_proxy_config(self) -> Union[Dict, ContainerServiceNetworkProfileKubeProxyConfig, None]:
+        """Obtain the value of kube_proxy_config.
+
+        :return: dictionary, ContainerServiceNetworkProfileKubeProxyConfig or None
+        """
+        # read the original value passed by the command
+        kube_proxy_config = None
+        kube_proxy_config_file_path = self.raw_param.get("kube_proxy_config")
+        # validate user input
+        if kube_proxy_config_file_path:
+            if not os.path.isfile(kube_proxy_config_file_path):
+                raise InvalidArgumentValueError(
+                    "{} is not valid file, or not accessible.".format(
+                        kube_proxy_config_file_path
+                    )
+                )
+            kube_proxy_config = get_file_json(kube_proxy_config_file_path)
+            if not isinstance(kube_proxy_config, dict):
+                raise InvalidArgumentValueError(
+                    "Error reading kube-proxy config from {}. "
+                    "Please see https://aka.ms/KubeProxyConfig for correct format.".format(
+                        kube_proxy_config_file_path
+                    )
+                )
+
+        # try to read the property value corresponding to the parameter from the `mc` object
+        if self.mc and self.mc.network_profile and self.mc.network_profile.kube_proxy_config is not None:
+            kube_proxy_config = self.mc.network_profile.kube_proxy_config
+
+        # this parameter does not need dynamic completion
+        # this parameter does not need validation
+        return kube_proxy_config
 
     def _get_enable_pod_security_policy(self, enable_validation: bool = False) -> bool:
         """Internal function to obtain the value of enable_pod_security_policy.
@@ -1861,7 +2028,6 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
 
     def _get_enable_node_restriction(self, enable_validation: bool = False) -> bool:
         """Internal function to obtain the value of enable_node_restriction.
-
         This function supports the option of enable_node_restriction. When enabled, if both enable_node_restriction and disable_node_restriction are
         specified, raise a MutuallyExclusiveArgumentError.
 
@@ -1878,6 +2044,64 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
                 )
 
         return enable_node_restriction
+
+    def _get_enable_azure_monitor_metrics(self, enable_validation: bool = False) -> bool:
+        """Internal function to obtain the value of enable_azure_monitor_metrics.
+        This function supports the option of enable_validation. When enabled, if both enable_azure_monitor_metrics and disable_azure_monitor_metrics are
+        specified, raise a MutuallyExclusiveArgumentError.
+
+        :return: bool
+        """
+        # print("_get_enable_azure_monitor_metrics being called...")
+        # Read the original value passed by the command.
+        enable_azure_monitor_metrics = self.raw_param.get("enable_azuremonitormetrics")
+        # In create mode, try to read the property value corresponding to the parameter from the `mc` object.
+        if self.decorator_mode == DecoratorMode.CREATE:
+            if (
+                self.mc and
+                self.mc.azure_monitor_profile and
+                self.mc.azure_monitor_profile.metrics
+            ):
+                enable_azure_monitor_metrics = self.mc.azure_monitor_profile.metrics.enabled
+        # This parameter does not need dynamic completion.
+        if enable_validation:
+            if enable_azure_monitor_metrics and self._get_disable_azure_monitor_metrics(False):
+                raise MutuallyExclusiveArgumentError(
+                    "Cannot specify --enable-azuremonitormetrics and --disable-azuremonitormetrics at the same time."
+                )
+            if enable_azure_monitor_metrics and not check_is_msi_cluster(self.mc):
+                raise RequiredArgumentMissingError(
+                    "--enable-azuremonitormetrics can only be specified for clusters with managed identity enabled"
+                )
+        return enable_azure_monitor_metrics
+
+    def get_enable_azure_monitor_metrics(self) -> bool:
+        """Obtain the value of enable_azure_monitor_metrics.
+        This function will verify the parameter by default. If both enable_azure_monitor_metrics and disable_azure_monitor_metrics are specified, raise a
+        MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        return self._get_enable_azure_monitor_metrics(enable_validation=True)
+
+    def _get_disable_azure_monitor_metrics(self, enable_validation: bool = False) -> bool:
+        """Internal function to obtain the value of disable_azure_monitor_metrics.
+        This function supports the option of enable_validation. When enabled, if both enable_azure_monitor_metrics and disable_azure_monitor_metrics are
+        specified, raise a MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        # Read the original value passed by the command.
+        disable_azure_monitor_metrics = self.raw_param.get("disable_azuremonitormetrics")
+        if disable_azure_monitor_metrics and self._get_enable_azure_monitor_metrics(False):
+            raise MutuallyExclusiveArgumentError("Cannot specify --enable-azuremonitormetrics and --disable-azuremonitormetrics at the same time.")
+        return disable_azure_monitor_metrics
+
+    def get_disable_azure_monitor_metrics(self) -> bool:
+        """Obtain the value of disable_azure_monitor_metrics.
+        This function will verify the parameter by default. If both enable_azure_monitor_metrics and disable_azure_monitor_metrics are specified, raise a
+        MutuallyExclusiveArgumentError.
+        :return: bool
+        """
+        return self._get_disable_azure_monitor_metrics(enable_validation=True)
 
     def get_enable_node_restriction(self) -> bool:
         """Obtain the value of enable_node_restriction.
@@ -1978,6 +2202,24 @@ class AKSPreviewManagedClusterContext(AKSManagedClusterContext):
         """
         return self._get_disable_vpa(enable_validation=True)
 
+    def get_ssh_key_value_for_update(self) -> Tuple[str, bool]:
+        """Obtain the value of ssh_key_value for "az aks update".
+
+        Note: no_ssh_key will not be decorated into the `mc` object.
+
+        If the user provides a string-like input for --ssh-key-value, the validator function "validate_ssh_key_for_update" will
+        check whether it is a file path, if so, read its content and return; if it is a valid public key, return it.
+        Otherwise, raise error.
+
+        :return: ssh_key_value of string type
+        """
+        # read the original value passed by the command
+        ssh_key_value = self.raw_param.get("ssh_key_value")
+
+        # this parameter does not need dynamic completion
+        # this parameter does not need validation
+        return ssh_key_value
+
 
 class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
     def __init__(
@@ -2065,6 +2307,9 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
             )
 
         network_profile.network_plugin_mode = self.context.get_network_plugin_mode()
+
+        if self.context.get_enable_cilium_dataplane():
+            network_profile.ebpf_dataplane = CONST_EBPF_DATAPLANE_CILIUM
 
         return mc
 
@@ -2162,7 +2407,9 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         """
         self._ensure_mc(mc)
 
-        mc.oidc_issuer_profile = self.context.get_oidc_issuer_profile()
+        oidc_issuer_profile = self.context.get_oidc_issuer_profile()
+        if oidc_issuer_profile is not None:
+            mc.oidc_issuer_profile = oidc_issuer_profile
 
         return mc
 
@@ -2321,7 +2568,7 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
 
         if self.context.get_enable_node_restriction():
             if mc.security_profile is None:
-                    mc.security_profile = self.models.ManagedClusterSecurityProfile()
+                mc.security_profile = self.models.ManagedClusterSecurityProfile()
             mc.security_profile.node_restriction = self.models.ManagedClusterSecurityProfileNodeRestriction(
                 enabled=True,
             )
@@ -2344,6 +2591,21 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
                 mc.workload_auto_scaler_profile.vertical_pod_autoscaler.enabled = True
         return mc
 
+    def set_up_kube_proxy_config(self, mc: ManagedCluster) -> ManagedCluster:
+        """Set up kube-proxy config for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if not mc.network_profile:
+            raise UnknownError(
+                "Unexpectedly get an empty network profile in the process of updating kube-proxy config."
+            )
+
+        mc.network_profile.kube_proxy_config = self.context.get_kube_proxy_config()
+        return mc
+
     def construct_mc_profile_preview(self, bypass_restore_defaults: bool = False) -> ManagedCluster:
         """The overall controller used to construct the default ManagedCluster profile.
 
@@ -2361,15 +2623,10 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         mc = self.set_up_pod_security_policy(mc)
         # set up pod identity profile
         mc = self.set_up_pod_identity_profile(mc)
-
-        # update workload identity & OIDC issuer settings
-        # NOTE: in current implementation, workload identity settings setup requires checking
-        #       previous OIDC issuer profile. However, the OIDC issuer settings setup will
-        #       overrides the previous OIDC issuer profile based on user input. Therefore, we have
-        #       to make sure the workload identity settings setup is done after OIDC issuer settings.
-        mc = self.set_up_workload_identity_profile(mc)
+        # set up oidc issuer profile, GA in 2.42.0
         mc = self.set_up_oidc_issuer_profile(mc)
-
+        # set up workload identity profile
+        mc = self.set_up_workload_identity_profile(mc)
         # set up azure keyvalut kms
         mc = self.set_up_azure_keyvault_kms(mc)
         # set up node restriction
@@ -2386,6 +2643,8 @@ class AKSPreviewManagedClusterCreateDecorator(AKSManagedClusterCreateDecorator):
         mc = self.set_up_workload_auto_scaler_profile(mc)
         # set up vpa
         mc = self.set_up_vpa(mc)
+        # set up kube-proxy config
+        mc = self.set_up_kube_proxy_config(mc)
 
         # DO NOT MOVE: keep this at the bottom, restore defaults
         mc = self._restore_defaults_in_mc(mc)
@@ -2467,6 +2726,40 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
                 )
                 raise RequiredArgumentMissingError(error_msg)
 
+    def update_outbound_type_in_network_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update outbound type of network profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        outboundType = self.context.get_outbound_type()
+        if outboundType:
+            mc.network_profile.outbound_type = outboundType
+        return mc
+
+    def update_nat_gateway_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update nat gateway profile for the ManagedCluster object.
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        if not mc.network_profile:
+            raise UnknownError(
+                "Unexpectedly get an empty network profile in the process of updating nat gateway profile."
+            )
+        outbound_type = self.context.get_outbound_type()
+        if outbound_type and outbound_type != CONST_OUTBOUND_TYPE_MANAGED_NAT_GATEWAY:
+            mc.network_profile.nat_gateway_profile = None
+        else:
+            mc.network_profile.nat_gateway_profile = _update_nat_gateway_profile(
+                self.context.get_nat_gateway_managed_outbound_ip_count(),
+                self.context.get_nat_gateway_idle_timeout(),
+                mc.network_profile.nat_gateway_profile,
+                models=self.models.nat_gateway_models,
+            )
+        return mc
+
     def update_load_balancer_profile(self, mc: ManagedCluster) -> ManagedCluster:
         """Update load balancer profile for the ManagedCluster object.
 
@@ -2480,21 +2773,24 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
             raise UnknownError(
                 "Unexpectedly get an empty network profile in the process of updating load balancer profile."
             )
-
-        # In the internal function "_update_load_balancer_profile", it will check whether the provided parameters
-        # have been assigned, and if there are any, the corresponding profile will be modified; otherwise, it will
-        # remain unchanged.
-        mc.network_profile.load_balancer_profile = _update_load_balancer_profile(
-            managed_outbound_ip_count=self.context.get_load_balancer_managed_outbound_ip_count(),
-            managed_outbound_ipv6_count=self.context.get_load_balancer_managed_outbound_ipv6_count(),
-            outbound_ips=self.context.get_load_balancer_outbound_ips(),
-            outbound_ip_prefixes=self.context.get_load_balancer_outbound_ip_prefixes(),
-            outbound_ports=self.context.get_load_balancer_outbound_ports(),
-            idle_timeout=self.context.get_load_balancer_idle_timeout(),
-            backend_pool_type=self.context.get_load_balancer_backend_pool_type(),
-            profile=mc.network_profile.load_balancer_profile,
-            models=self.models.load_balancer_models,
-        )
+        outbound_type = self.context.get_outbound_type()
+        if outbound_type and outbound_type != CONST_OUTBOUND_TYPE_LOAD_BALANCER:
+            mc.network_profile.load_balancer_profile = None
+        else:
+            # In the internal function "_update_load_balancer_profile", it will check whether the provided parameters
+            # have been assigned, and if there are any, the corresponding profile will be modified; otherwise, it will
+            # remain unchanged.
+            mc.network_profile.load_balancer_profile = _update_load_balancer_profile(
+                managed_outbound_ip_count=self.context.get_load_balancer_managed_outbound_ip_count(),
+                managed_outbound_ipv6_count=self.context.get_load_balancer_managed_outbound_ipv6_count(),
+                outbound_ips=self.context.get_load_balancer_outbound_ips(),
+                outbound_ip_prefixes=self.context.get_load_balancer_outbound_ip_prefixes(),
+                outbound_ports=self.context.get_load_balancer_outbound_ports(),
+                idle_timeout=self.context.get_load_balancer_idle_timeout(),
+                backend_pool_type=self.context.get_load_balancer_backend_pool_type(),
+                profile=mc.network_profile.load_balancer_profile,
+                models=self.models.load_balancer_models,
+            )
         return mc
 
     def update_api_server_access_profile(self, mc: ManagedCluster) -> ManagedCluster:
@@ -2601,8 +2897,9 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         :return: the ManagedCluster object
         """
         self._ensure_mc(mc)
-
-        mc.oidc_issuer_profile = self.context.get_oidc_issuer_profile()
+        oidc_issuer_profile = self.context.get_oidc_issuer_profile()
+        if oidc_issuer_profile is not None:
+            mc.oidc_issuer_profile = oidc_issuer_profile
 
         return mc
 
@@ -2756,6 +3053,47 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
 
         return mc
 
+    def update_azure_monitor_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update azure monitor profile for the ManagedCluster object.
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        # read the original value passed by the command
+        ksm_metric_labels_allow_list = self.context.raw_param.get("ksm_metric_labels_allow_list")
+        ksm_metric_annotations_allow_list = self.context.raw_param.get("ksm_metric_annotations_allow_list")
+
+        if ksm_metric_labels_allow_list is None:
+            ksm_metric_labels_allow_list = ""
+        if ksm_metric_annotations_allow_list is None:
+            ksm_metric_annotations_allow_list = ""
+
+        if self.context.get_enable_azure_monitor_metrics():
+            if mc.azure_monitor_profile is None:
+                mc.azure_monitor_profile = self.models.ManagedClusterAzureMonitorProfile()
+            mc.azure_monitor_profile.metrics = self.models.ManagedClusterAzureMonitorProfileMetrics(enabled=True)
+            mc.azure_monitor_profile.metrics.kube_state_metrics = self.models.ManagedClusterAzureMonitorProfileKubeStateMetrics(
+                metric_labels_allowlist=str(ksm_metric_labels_allow_list),
+                metric_annotations_allow_list=str(ksm_metric_annotations_allow_list))
+
+        if self.context.get_disable_azure_monitor_metrics():
+            if mc.azure_monitor_profile is None:
+                mc.azure_monitor_profile = self.models.ManagedClusterAzureMonitorProfile()
+            mc.azure_monitor_profile.metrics = self.models.ManagedClusterAzureMonitorProfileMetrics(enabled=False)
+
+        if (self.context.raw_param.get("enable_azuremonitormetrics") or self.context.raw_param.get("disable_azuremonitormetrics")):
+            ensure_azure_monitor_profile_prerequisites(
+                self.cmd,
+                self.client,
+                self.context.get_subscription_id(),
+                self.context.get_resource_group_name(),
+                self.context.get_name(),
+                self.context.get_location(),
+                self.__raw_parameters,
+                self.context.get_disable_azure_monitor_metrics())
+
+        return mc
+
     def update_node_restriction(self, mc: ManagedCluster) -> ManagedCluster:
         """Update security profile nodeRestriction for the ManagedCluster object.
 
@@ -2825,6 +3163,28 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
                 source_resource_id=snapshot_id
             )
             mc.creation_data = creation_data
+
+        return mc
+
+    def update_linux_profile(self, mc: ManagedCluster) -> ManagedCluster:
+        """Update Linux profile for the ManagedCluster object.
+
+        :return: the ManagedCluster object
+        """
+        self._ensure_mc(mc)
+
+        ssh_key_value = self.context.get_ssh_key_value_for_update()
+
+        if ssh_key_value:
+            if mc.linux_profile is None:
+                raise InvalidArgumentValueError("Updating the cluster with no ssh key set at creation is not supported")
+            mc.linux_profile.ssh = self.models.ContainerServiceSshConfiguration(
+                public_keys=[
+                    self.models.ContainerServiceSshPublicKey(
+                        key_data=ssh_key_value
+                    )
+                ]
+            )
         return mc
 
     def update_mc_profile_preview(self) -> ManagedCluster:
@@ -2844,15 +3204,10 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_pod_security_policy(mc)
         # update pod identity profile
         mc = self.update_pod_identity_profile(mc)
-
-        # update workload identity & OIDC issuer settings
-        # NOTE: in current implementation, workload identity settings setup requires checking
-        #       previous OIDC issuer profile. However, the OIDC issuer settings setup will
-        #       overrides the previous OIDC issuer profile based on user input. Therefore, we have
-        #       to make sure the workload identity settings setup is done after OIDC issuer settings.
-        mc = self.update_workload_identity_profile(mc)
+        # update oidc issure profile, GA in 2.42.0
         mc = self.update_oidc_issuer_profile(mc)
-
+        # update workload identity profile
+        mc = self.update_workload_identity_profile(mc)
         # update azure keyvalut kms
         mc = self.update_azure_keyvault_kms(mc)
         # update node restriction
@@ -2863,8 +3218,15 @@ class AKSPreviewManagedClusterUpdateDecorator(AKSManagedClusterUpdateDecorator):
         mc = self.update_storage_profile(mc)
         # update workload auto scaler profile
         mc = self.update_workload_auto_scaler_profile(mc)
+        # update azure monitor metrics profile
+        mc = self.update_azure_monitor_profile(mc)
         # update vpa
         mc = self.update_vpa(mc)
         # update creation data
         mc = self.update_creation_data(mc)
+        # update linux profile
+        mc = self.update_linux_profile(mc)
+        # update outbound type
+        mc = self.update_outbound_type_in_network_profile(mc)
+
         return mc