+++ src/spring/azext_spring/spring_instance.py
@@ -11,13 +11,18 @@ from .custom import (_warn_enable_java_agent, _update_application_insights_asc_c
 from ._build_service import _update_default_build_agent_pool
 from .buildpack_binding import create_default_buildpack_binding_for_application_insights
 from ._tanzu_component import (create_application_configuration_service,
+                               create_application_live_view,
+                               create_dev_tool_portal,
                                create_service_registry,
                                create_gateway,
-                               create_api_portal)
-
+                               create_api_portal,
+                               create_application_accelerator)
 
 from ._validators import (_parse_sku_name, validate_instance_not_existed)
+from azure.cli.core.commands import LongRunningOperation
 from knack.log import get_logger
+from ._marketplace import _spring_list_marketplace_plan
+from ._constant import (MARKETPLACE_OFFER_ID, MARKETPLACE_PUBLISHER_ID)
 
 logger = get_logger(__name__)
 
@@ -39,7 +44,6 @@ class DefaultSpringCloud:
     def before_create(self, **kwargs):
         _warn_enable_java_agent(**kwargs)
         validate_instance_not_existed(self.client,
-                                      self.resource_group,
                                       self.name,
                                       self.location)
 
@@ -62,6 +66,7 @@ class DefaultSpringCloud:
                        sku=None,
                        tags=None,
                        ingress_read_timeout=None,
+                       marketplace_plan_id=None,
                        **_):
         properties = models.ClusterResourceProperties(
             zone_redundant=zone_redundant
@@ -73,6 +78,13 @@ class DefaultSpringCloud:
             )
         else:
             properties.vnet_addons = None
+
+        if marketplace_plan_id:
+            properties.marketplace_resource = models.MarketplaceResource(
+                plan=marketplace_plan_id,
+                product=MARKETPLACE_OFFER_ID,
+                publisher=MARKETPLACE_PUBLISHER_ID
+            )
 
         if service_runtime_subnet or app_subnet or reserved_cidr_range:
             properties.network_profile = models.NetworkProfile(
@@ -95,14 +107,12 @@ class DefaultSpringCloud:
         poller = self.client.services.begin_create_or_update(
             self.resource_group, self.name, resource)
         logger.warning(" - Creating Service ..")
-        wait_till_end(self.cmd, poller)
-        return poller
+        return LongRunningOperation(self.cmd.cli_ctx)(poller)
 
 
 class EnterpriseSpringCloud(DefaultSpringCloud):
     def before_create(self, **_):
         validate_instance_not_existed(self.client,
-                                      self.resource_group,
                                       self.name,
                                       self.location)
 
@@ -113,9 +123,12 @@ class EnterpriseSpringCloud(DefaultSpringCloud):
                 self.cmd, self.client, self.resource_group, self.name, kwargs['build_pool_size']),
             _enable_app_insights(self.cmd, self.client, self.resource_group, self.name, self.location, **kwargs),
             create_application_configuration_service(self.cmd, self.client, self.resource_group, self.name, **kwargs),
+            create_application_live_view(self.cmd, self.client, self.resource_group, self.name, **kwargs),
+            create_dev_tool_portal(self.cmd, self.client, self.resource_group, self.name, **kwargs),
             create_service_registry(self.cmd, self.client, self.resource_group, self.name, **kwargs),
             create_gateway(self.cmd, self.client, self.resource_group, self.name, **kwargs),
-            create_api_portal(self.cmd, self.client, self.resource_group, self.name, **kwargs)
+            create_api_portal(self.cmd, self.client, self.resource_group, self.name, **kwargs),
+            create_application_accelerator(self.cmd, self.client, self.resource_group, self.name, **kwargs)
         ]
         pollers = [x for x in pollers if x]
         if not no_wait:
@@ -147,13 +160,16 @@ def spring_create(cmd, client, resource_group, name,
                   zone_redundant=False,
                   build_pool_size=None,
                   enable_application_configuration_service=False,
+                  enable_application_live_view=False,
                   enable_service_registry=False,
                   enable_gateway=False,
                   gateway_instance_count=None,
                   enable_api_portal=False,
                   api_portal_instance_count=None,
+                  enable_application_accelerator=False,
                   enable_log_stream_public_endpoint=None,
                   ingress_read_timeout=None,
+                  marketplace_plan_id=None,
                   no_wait=False):
     """
     Because Standard/Basic tier vs. Enterprise tier creation are very different. Here routes the command to different
@@ -178,12 +194,15 @@ def spring_create(cmd, client, resource_group, name,
         'zone_redundant': zone_redundant,
         'build_pool_size': build_pool_size,
         'enable_application_configuration_service': enable_application_configuration_service,
+        'enable_application_live_view': enable_application_live_view,
         'enable_service_registry': enable_service_registry,
         'enable_gateway': enable_gateway,
         'gateway_instance_count': gateway_instance_count,
         'enable_api_portal': enable_api_portal,
         'api_portal_instance_count': api_portal_instance_count,
+        'enable_application_accelerator': enable_application_accelerator,
         'enable_log_stream_public_endpoint': enable_log_stream_public_endpoint,
+        'marketplace_plan_id': marketplace_plan_id,
         'no_wait': no_wait
     }
 
@@ -199,3 +218,7 @@ def _enable_app_insights(cmd, client, resource_group, name, location, app_insigh
     return create_default_buildpack_binding_for_application_insights(cmd, client, resource_group, name,
                                                                      location, app_insights_key, app_insights,
                                                                      sampling_rate)
+
+
+def spring_list_marketplace_plan(cmd, client):
+    return _spring_list_marketplace_plan(cmd, client)